requirements.txt
Django==4.0.1
python-engineio==4.3.1
python-socketio==5.5.1
sqlparse==0.4.1
whitenoise==5.3.0
gevent==21.8.0
gevent-websocket==0.10.1
mysqlclient==2.0.3
psycopg2==2.9.3
python-decouple==3.5